Update running CEL validation in Worker thread
